# usf/src/main/java/uol/compass/project/usf/model/entities/ResourceEntity.java
package uol.compass.project.usf.model.entities;

import lombok.*;
import uol.compass.project.usf.model.enums.EnumCategoryResource;

import javax.persistence.*;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity(name = "resource")
@Table(name = "resource")
public class ResourceEntity {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(name = "name")
    private String name;

    @Column(name = "description")
    private String description;

    @Column(name = "category")
    @Enumerated(EnumType.STRING)
    private EnumCategoryResource category;

}