# usf/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Swagger USF API - OpenAPI 0.0.1
  description: |-
    A USF (Unidade de Saúde da Família) nada mais é que uma unidade pública de saúde destinada
    a realizar atenção contínua nas especialidades básicas, com uma equipe multiprofissional
    habiitada para desenvolver as atividades de promoção, proteção e recuperação, características
    do nível primário de atenção. Após conversa com funcionários dessas unidades notamos que há uma
    complexidade na gestão desse ambiente, pois precisam fazer a comunicação com outras USFs, seja para
    realizar pedidos de recursos ou para gestão de profissionais, de maneira manual através de pranchetas 
    e documentos por escrito. Fazendo o sistema desses ambientes vitais para o cuidado das comunidades brasileiras 
    ficam propensos a erros e lentidão no atendimento.

    Objetivando desenvolvimento para esse serviço, desenvolvemos uma API para automatizar esses processos defasados.


    Alguns links que podem ser interessantes:
    - [Repositório da USF API](https://github.com/solanokruger/usf-project)
    - [Toda documentação do projeto](https://docs.google.com/document/d/1Rgm3JiFKQBkO_tDVDDw1Mr-FGm9hP7MV-BGx2f4m3HE/edit?usp=sharing)
  termsOfService: http://swagger.io/terms/
  contact:
    email: luis.gustavo@compasso.com.br
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: API local
...
paths:
...
  /resource:
    post:
      security:
        - bearerAuth: []
      tags:
        - Resource
      summary: Adiciona um novo recurso
      description: Adiciona um novo recurso
      operationId: addResource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceRequestDTO"
        required: true
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"

    get:
      security:
        - bearerAuth: []
      tags:
        - Resource
      summary: Retorna todos os Recursos
      description: Retorna todos os Recursos
      operationId: getResource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponseParameters"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
  /resource/{resourceId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Resource
      summary: Procura um recurso pelo ID
      description: Retorna apenas um recurso
      operationId: getResourceById
      parameters:
        - name: resourceId
          in: path
          description: ID do recurso que vai ser retornado
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: RESOURCE NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    put:
      security:
        - bearerAuth: []
      tags:
        - Resource
      summary: Atualiza um recurso
      description: Atualiza um recurso existente pelo id
      operationId: updateResource
      parameters:
        - name: resourceId
          in: path
          description: ID do recurso que vai ser retornado
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Atualiza um recurso existente pelo id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: RESOURCE NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Resource
      summary: Remove um recurso
      description: Remove um recurso pelo seu ID
      operationId: deleteResource
      parameters:
        - name: resourceId
          in: path
          description: ID do recurso que será removido
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: RESOURCE NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
  /inventory:
    post:
      security:
        - bearerAuth: []
      tags:
        - Inventory
      summary: Adiciona um novo inventário
      description: Adiciona um novo inventário
      operationId: addInventory
      requestBody:
        description: Adiciona um novo inventário
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryRequestDTO"
        required: true
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"

    get:
      security:
        - bearerAuth: []
      tags:
        - Inventory
      summary: Lista todos os inventários
      description: Lista todos os inventários
      operationId: getInventory
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponseParameters"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
  /inventory/{inventoryId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Inventory
      summary: Procura um inventário pelo ID
      description: Retorna somente um inventário
      operationId: getInventoryById
      parameters:
        - name: inventoryId
          in: path
          description: ID do inventário para procura
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: INVENTORY NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    put:
      security:
        - bearerAuth: []
      tags:
        - Inventory
      summary: Atualiza um inventário existente
      description: Atualiza um inventário existente pelo ID.
      operationId: updateInventory
      parameters:
        - name: inventoryId
          in: path
          description: ID do inventário para procura
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Atualiza um inventário pelo ID.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: INVENTORY NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Inventory
      summary: Remove um inventário
      description: Remove um inventário pelo seu ID
      operationId: deleteInventory
      parameters:
        - name: inventoryId
          in: path
          description: ID do inventário que será removido
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: INVENTORY NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
...
components:
  schemas:
...
    InventoryRequestDTO:
      type: object
      properties:
        idResource:
          type: integer
          format: int64
        idUsf:
          type: integer
          format: int64
        amount:
          type: integer
      example:
        idResource: 1
        idUsf: 1
        amount: 50
    ResourceRequestDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
      example:
        name: Gaze
        description: Usada pra fazer curativos em machucados na pele
        category: EQUIPAMENTO
...
    ResourceEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [EQUIPAMENTO, MEDICAMENTO]
      example:
        id: 1
        name: ÁLCOOL
        description: Álcool 70% antisseptico
        category: EQUIPAMENTO
    UsfEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
      example:
        id: 1
        name: Unidade de Saúde da Família da Avenida 2 de Julho
        address: Avenida 2 de Julho, número 15
    InventoryResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        resource:
          $ref: "#/components/schemas/ResourceEntity"
        usf:
          $ref: "#/components/schemas/UsfEntity"
        amount:
          type: integer
      example:
        id: 1
        amount: 67
    InventoryResponseParameters:
      type: object
      properties:
        numberOfElements:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/InventoryResponseDTO"
      example:
        numberOfElements: 10
        totalElements: 10
        totalPages: 1
    ResourceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [EQUIPAMENTO, MEDICAMENTO]
      example:
        id: 1
        name: Gaze
        description: Usada pra fazer curativos em machucados na pele
        category: EQUIPAMENTO
    ResourceResponseParameters:
      type: object
      properties:
        numberOfElements:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ResourceResponseDTO"
      example:
        numberOfElements: 10
        totalElements: 10
        totalPages: 1
...
    MethodArgumentNotValidException:
      type: object
      properties:
        details:
          type: string
        errorCode:
          type: string
        httpStatus:
          type: string
    NotFoundException:
      type: object
      properties:
        details:
          type: string
        errorCode:
          type: string
        httpStatus:
          type: string
...
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT